+++ inference-engine/tests/ngraph_functions/src/split.cpp
@@ -12,9 +12,9 @@ namespace builder {
 std::shared_ptr<ngraph::Node> makeSplit(const ngraph::Output<Node> &in,
                                         const element::Type &type,
                                         size_t numSplits,
-                                        size_t axis) {
+                                        int64_t axis) {
     auto splitAxisOp = std::make_shared<ngraph::opset1::Constant>(ngraph::element::Type_t::i64, ngraph::Shape{},
-                                                                  std::vector<uint64_t>{axis});
+                                                                  std::vector<int64_t>{axis});
     auto splitNode = std::make_shared<ngraph::opset1::Split>(in, splitAxisOp, numSplits);
     return splitNode;
 }
